pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-runner-workspace"
version = "0.1.0"
description = "Agent Runner monorepo: orchestrator, registry, and evaluation harness"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "httpx>=0.27",
  "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "import-linter>=2.0",
  "ruff>=0.6",
]
observability = [
  "langfuse>=2.0",
]

[project.scripts]
agent-runner = "agent_runner.cli.headless:main"
agent-runner-interactive = "agent_runner.cli.interactive:main"
agent-runner-general = "agent_runner.cli.general:main"
agent-runner-harness = "agent_runner_harness.cli.main:main"
agent-runner-registry = "agent_runner_registry.cli:main"

[tool.setuptools]
package-dir = {"" = "packages/shared"}

[tool.setuptools.packages.find]
where = [
  "packages/shared",
  "packages/runner",
  "packages/registry",
  "packages/harness",
]

[tool.pytest.ini_options]
testpaths = ["tests", "packages/runner/tests", "packages/shared/tests", "packages/registry/tests", "packages/harness/tests"]
pythonpath = ["packages/shared", "packages/runner", "packages/registry", "packages/harness"]

[tool.importlinter]
root_packages = ["agent_runner", "agent_runner_shared", "agent_runner_registry", "agent_runner_harness"]

[[tool.importlinter.contracts]]
name = "Shared has no dependencies on other packages"
type = "forbidden"
source_modules = ["agent_runner_shared"]
forbidden_modules = ["agent_runner", "agent_runner_registry", "agent_runner_harness"]

[[tool.importlinter.contracts]]
name = "Runner does not depend on harness or registry internals"
type = "forbidden"
source_modules = ["agent_runner"]
forbidden_modules = ["agent_runner_harness"]

[[tool.importlinter.contracts]]
name = "Harness does not depend on runner internals"
type = "forbidden"
source_modules = ["agent_runner_harness"]
forbidden_modules = ["agent_runner"]
